requirements.txt
gunicorn==0.17.4
Flask==0.10.1
Jinja2==2.7
MarkupSafe==0.18
PyYAML==3.10
Werkzeug==0.9.1
argparse==1.2.1
httplib2==0.7.7
itsdangerous==0.21
numpy==1.7.1
oauth2==1.5.211
python-twitter==1.0
simplejson==3.3.0
tweepy==2.0
wsgiref==0.1.2
nltk==2.0.4
pylibmc==1.2.3
boto==2.9.6
